+++ CHANGELOG.md
@@ -6,11 +6,12 @@ Version 2.5-SNAPSHOT
 
 ADDED:
  
-- added new test-dependency junit-jupiter (junit 5) in version 5.6.2
+- added new test-dependency junit-jupiter (junit 5) in version 5.8.1
 
 CHANGED:
 
-- update gradle to new version 6.5.1
+- update gradle to new version 7.2
+- changed to new package io.github.astrapi69
 - remove of test-dependency testng 
 
 Version 2.4